requirements.txt
Django==1.8
Pillow==3.1.1
qiniu==7.0.5
pytz==2015.7
python-memcached==1.57
django-smtp-ssl==1.0
